# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Winding tree API",
    "description": "Node js API for Winding Tree",
    "contact": {
      "name": "Winding tree",
      "url": "https://windingtree.com/"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
...
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/",
...
  "paths": {
...
    "/hotels": {
      "get": {
        "tags": [ "Hotel" ],
        "summary": "Retrieves all hotels.",
        "description": "Retrieves hotels for the account.",
        "parameters": [
          {
            "name": "password",
            "in": "body",
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password. This will no longer be the current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "schema": {
              "$ref": "#/definitions/Hotels"
            }
          },
          "400": {
            "description": "Can not get.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": [ "Hotel" ],
        "summary": "Create hotel",
        "description": "Create a new hotel with name and description.",
        "parameters": [
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password",
                "name",
                "description"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "name": {
                  "description": "Hotel's name.",
                  "type": "string"
                },
                "description": {
                  "description": "Hotel's description.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully created.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}":{
      "get":{
        "tags": [ "Hotel" ],
        "summary": "Get Hotel.",
        "description": "Get Hotel by contract address.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "name": "password",
            "in": "body",
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "schema": {
              "$ref": "#/definitions/Hotels"
            }
          },
          "400": {
            "description": "Can not get.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [ "Hotel" ],
        "summary": "Update hotel info",
        "description": "Update hotel name and description.",
        "parameters": [
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password",
                "name",
                "description"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "name": {
                  "description": "Hotel's name.",
                  "type": "string"
                },
                "description": {
                  "description": "Hotel's description.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully updated.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete":{
        "tags": [ "Hotel" ],
        "summary": "Remove Hotel.",
        "description": "Remove Hotel by contract address.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "name": "password",
            "in": "body",
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Bad parameters.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/address": {
      "put": {
        "tags": [ "Hotel" ],
        "summary": "Hotel's address.",
        "description": "Change hotel's address",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password",
                "lineOne",
                "lineTwo",
                "zipCode",
                "country"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "lineOne": {
                  "description": "street number , street name.",
                  "type": "string"
                },
                "lineTwo": {
                  "description": "street number , street name.",
                  "type": "string"
                },
                "zipCode": {
                  "description": "Hotel's zip code.",
                  "type": "string"
                },
                "country": {
                  "description": "Hotel's country.",
                  "type": "string"
                 }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items" : {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/hotels/{hotelAddress}/location": {
      "put": {
        "tags": [ "Hotel" ],
        "summary": "Hotel's location.",
        "description": "Change hotel's location",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel location."
          },
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password",
                "timezon",
                "latitude",
                "longitude"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "latitude": {
                  "description": "Hotels's latitude.",
                  "type": "string"
                },
                "longitude": {
                  "description": "Hotels's longitude.",
                  "type": "string"
                },
                "timezone": {
                  "description": "Hotel's timezone.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/images": {
      "get": {
        "tags": [ "Hotel" ],
        "summary": "Hotel's images.",
        "description": "Retrieves hotel's",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items" : {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": [ "Hotel" ],
        "summary": "Add image to hotel",
        "description": "Add an image of the hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password",
                "url"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "url": {
                  "description": "Picture's url.",
                  "type": "string"
                 }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully created.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/hotels/{hotelAddress}/images/{id}": {
        "delete": {
          "tags": [ "Hotel" ],
          "summary": "remove image from hotel.",
          "description": "Remove image from hotel by id.",
          "parameters": [
            {
              "in": "path",
              "name": "address",
              "type": "string",
              "required": true,
              "description": "Hotel address."
            },
            {
              "in": "path",
              "name": "id",
              "type": "integer",
              "required": true,
              "description": "Image id."
            },
            {
              "in": "body",
              "name": "password",
              "required": true,
              "description": "Password.",
              "schema": {
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "description": "Current password.",
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Sucefully removed.",
              "schema": {
                "$ref": "#/definitions/txHash"
              }
            },
            "400": {
              "description": "Invalid input",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
    },
...
  },
  "definitions": {
    "txHash": {
      "title": "Transaction hash",
      "description": "transaction hash for the operation.",
      "properties": {
        "txHash": {
          "type": "string",
          "description": "Transaction hash."
        }
      }
    },
    "Error": {
      "title": "Error",
      "description": "Default schema for errors returned by API.",
      "properties": {
        "code": {
          "type": "string",
          "description": "Computer-readable error. Use this for comparison."
        },
        "short": {
          "type": "string",
          "description": "Human-readable error with a short description of the error."
        },
        "long": {
          "type": "string",
          "description": "Verbose explanation of what happen."
        }
      }
    },
    "Hotels": {
      "title": "Hotels",
      "description": "Hotels list.",
      "type": "array",
      "items": {
        "$ref": "#/definitions/Hotel"
      }
    },
    "Hotel": {
      "title": "Hotel",
      "description": "Hotel properties",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "manager": {
          "type": "string"
        },
        "lineOne": {
          "type": "string"
        },
        "lineTwo": {
          "type": "string"
        },
        "zip": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "created": {
          "type": "integer"
        },
        "timezone": {
          "type": "string"
        },
        "latitude": {
          "type": "number"
        },
        "longitude": {
          "type": "number"
        },
        "waitConfirmation": {
          "type": "boolean"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "unitTypeNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "unitTypes": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "units": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "unitAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}